Move system middlewares to App\Extensions namespace
