+++ bootstrap/app.php
@@ -6,9 +6,9 @@ use App\Extensions\LaravelContext\Middlewares\UpdateContext;
 use App\Facades\System\AppException;
 use App\Http\Middleware\HandleAppearance;
 use App\Http\Middleware\HandleInertiaRequests;
-use App\Http\Middleware\System\AddSecurityHeaders;
-use App\Http\Middleware\System\SetUserLocale;
-use App\Http\Middleware\System\TerminatingMiddleware;
+use App\System\Middlewares\AddSecurityHeaders;
+use App\System\Middlewares\SetUserLocale;
+use App\System\Middlewares\TerminatingMiddleware;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
